Clean up files, rename apiCalls to utilities
