+++ ui/admin/package.json
@@ -9,7 +9,7 @@
     "run": "node server",
     "dev-build": "webpack --mode development",
     "dev-run": "nodemon server",
-    "dev": "webpack-dev-server --watch"
+    "dev": "concurrently \"webpack --watch\" \"nodemon server\""
   },
   "keywords": [],
   "author": "",
@@ -21,7 +21,8 @@
     "babel-loader": "^8.0.6",
     "webpack": "^4.41.5",
     "webpack-cli": "^3.3.10",
-    "webpack-dev-server": "^3.10.1"
+    "webpack-dev-server": "^3.10.1",
+    "concurrently": "^5.0.2"
   },
   "dependencies": {
     "axios": "^0.19.2",
