+++ ui/admin/src/components/Admin.jsx
@@ -1,16 +1,7 @@
 import React from "react";
-import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
-
-import ProductsRouter from "./products/ProductsRouter";
 
 const Admin = ({ match }) => {
-  return (
-    <Router>
-      <Switch>
-        <Route path={`${match.url}/products`} component={ProductsRouter} />
-      </Switch>
-    </Router>
-  );
+  return <div></div>;
 };
 
 export default Admin;
