+++ ui/admin/src/components/products/productList/ConfirmDelete.jsx
@@ -1,9 +1,7 @@
 import React from "react";
-import { deleteItem } from "../../../apiCalls";
 import PropTypes from "prop-types";
 
 const ConfirmDelete = ({ product, deleteProduct }) => {
-  console.log("in confirm delete", product);
   return (
     <div
       className="modal fade"
