+++ ui/admin/src/components/products/productList/ProductList.jsx
@@ -1,21 +1,18 @@
-import React, { useEffect } from "react";
+import React from "react";
 import PropTypes from "prop-types";
 
 import ListItem from "./ListItem";
 
 const ProductList = ({ products, updateProducts }) => {
-  useEffect(() => {}, []);
   return (
     <ul className="list-group mt-3">
-      {products &&
-        products &&
-        products.map(product => (
-          <ListItem
-            product={product}
-            updateProducts={updateProducts}
-            key={product._id}
-          />
-        ))}
+      {products.map(product => (
+        <ListItem
+          product={product}
+          updateProducts={updateProducts}
+          key={product._id}
+        />
+      ))}
     </ul>
   );
 };
