+++ ui/proxy/package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "nodemon --watch"
+    "dev": "concurrently \"webpack --watch \" \"nodemon server\""
   },
   "keywords": [],
   "author": "",
@@ -18,6 +18,7 @@
     "@babel/core": "^7.8.3",
     "@babel/preset-env": "^7.8.3",
     "babel-loader": "^8.0.6",
+    "concurrently": "^5.0.2",
     "webpack": "^4.41.5",
     "webpack-cli": "^3.3.10"
   }
