--- ui/proxy/server/admin.js
@@ -1,8 +1,0 @@
-const express = require("express");
-const path = require("path");
-
-const router = express.Router();
-
-module.exports = router.get("*", (req, res) => {
-  res.sendFile(path.join(__dirname, "../dist/index.html"));
-});
