+++ ui/proxy/webpack.config.js
@@ -4,7 +4,7 @@ module.exports = {
   entry: "./index.js",
   output: {
     path: path.join(__dirname, "dist"),
-    filename: "index.js"
+    filename: "bundle.js"
   },
   module: {
     rules: [
